Origin code

+++ db/migrate/20200823104440_add_password_to_users.rb
@@ -1,7 +1,12 @@
 class AddPasswordToUsers < ActiveRecord::Migration[6.0]
   def change
-    remove_column :users, :flight_id
-    remove_column :users, :noofseat
+    if column_exists? :users, :flight_id 
+      remove_column :users, :flight_id
+    end
+    if column_exists? :users, :noofseat
+      remove_column :users, :noofseat
+    end
+
     add_column :users, :password_digest, :string
   end
 end
